Help Center: refactor contact form
