+++ packages/help-center/src/components/help-center-contact-form.tsx
@@ -4,11 +4,8 @@
  */
 import { recordTracksEvent } from '@automattic/calypso-analytics';
 import config from '@automattic/calypso-config';
-import { getPlan, getPlanTermLabel } from '@automattic/calypso-products';
 import { Button, FormInputValidation, Popover } from '@automattic/components';
 import {
-	useSubmitTicketMutation,
-	useSubmitForumsMutation,
 	useSiteAnalysis,
 	useUserSites,
 	AnalysisReport,
@@ -17,29 +14,27 @@ import {
 	HelpCenterSite,
 } from '@automattic/data-stores';
 import { useLocale } from '@automattic/i18n-utils';
-import { SitePickerDropDown, SitePickerSite } from '@automattic/site-picker';
 import { TextControl, CheckboxControl, Tip } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import { Icon, info } from '@wordpress/icons';
-import React, { useEffect, useRef, useState } from 'react';
-import { useQueryClient } from 'react-query';
+import { useEffect, useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { useHistory, useLocation } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import { useDebounce } from 'use-debounce';
-import { isWcMobileApp } from 'calypso/lib/mobile-app';
-import { getQueryArgs } from 'calypso/lib/query-args';
-import { getCurrentUserEmail, getCurrentUserId } from 'calypso/state/current-user/selectors';
+import { getCurrentUserId } from 'calypso/state/current-user/selectors';
 import { getSectionName } from 'calypso/state/ui/selectors';
 /**
  * Internal Dependencies
  */
+import { useContactFormCTA } from '../hooks/use-contact-form-cta';
+import { useContactFormTitle } from '../hooks/use-contact-form-title';
 import { HELP_CENTER_STORE } from '../stores';
 import { getSupportVariationFromMode } from '../support-variations';
-import { SitePicker } from '../types';
+import { Mode } from '../types';
 import { BackButton } from './back-button';
-import { HelpCenterOwnershipNotice } from './help-center-notice';
 import { SibylArticles } from './help-center-sibyl-articles';
+import { HelpCenterSitePicker } from './help-center-site-picker';
 import './help-center-contact-form.scss';
 
 export const SITE_STORE = 'automattic/site';
@@ -58,70 +53,6 @@ const fakeFaces = [
 	'tony',
 ].map( ( name ) => `https://s0.wp.com/i/support-engineers/${ name }.jpg` );
 const randomTwoFaces = fakeFaces.sort( () => Math.random() - 0.5 ).slice( 0, 2 );
-
-const HelpCenterSitePicker: React.FC< SitePicker > = ( {
-	onSelect,
-	currentSite,
-	siteId,
-	enabled,
-} ) => {
-	const otherSite = {
-		name: __( 'Other site', __i18n_text_domain__ ),
-		ID: 0,
-		logo: { id: '', sizes: [] as never[], url: '' },
-		URL: '',
-	} as const;
-
-	function pickSite( ID: number | string ) {
-		onSelect( ID );
-	}
-
-	const options: ( SitePickerSite | undefined )[] = [ currentSite, otherSite ];
-
-	return (
-		<SitePickerDropDown
-			enabled={ enabled }
-			onPickSite={ pickSite }
-			options={ options }
-			siteId={ siteId }
-		/>
-	);
-};
-
-function useFormTitles( mode: Mode ): {
-	formTitle: string;
-	formSubtitle?: string;
-	trayText?: string;
-	formDisclaimer?: string;
-	buttonLabel: string;
-	buttonSubmittingLabel: string;
-	buttonLoadingLabel?: string;
-} {
-	return {
-		CHAT: {
-			formTitle: __( 'Start live chat', __i18n_text_domain__ ),
-			trayText: __( 'Our WordPress experts will be with you right away', __i18n_text_domain__ ),
-			buttonLabel: __( 'Chat with us', __i18n_text_domain__ ),
-			buttonSubmittingLabel: __( 'Connecting to chat', __i18n_text_domain__ ),
-		},
-		EMAIL: {
-			formTitle: __( 'Send us an email', __i18n_text_domain__ ),
-			trayText: __( 'Our WordPress experts will get back to you soon', __i18n_text_domain__ ),
-			buttonLabel: __( 'Email us', __i18n_text_domain__ ),
-			buttonSubmittingLabel: __( 'Sending email', __i18n_text_domain__ ),
-		},
-		FORUM: {
-			formTitle: __( 'Ask in our community forums', __i18n_text_domain__ ),
-			formDisclaimer: __(
-				'Please do not provide financial or contact information when submitting this form.',
-				__i18n_text_domain__
-			),
-			buttonLabel: __( 'Ask in the forums', __i18n_text_domain__ ),
-			buttonSubmittingLabel: __( 'Posting in the forums', __i18n_text_domain__ ),
-			buttonLoadingLabel: __( 'Analyzing site…', __i18n_text_domain__ ),
-		},
-	}[ mode ];
-}
 
 const getSupportedLanguages = ( supportType: string, locale: string ) => {
 	const isLiveChatLanguageSupported = (
@@ -143,25 +74,16 @@ const getSupportedLanguages = ( supportType: string, locale: string ) => {
 	}
 };
 
-type Mode = 'CHAT' | 'EMAIL' | 'FORUM';
-
 export const HelpCenterContactForm = () => {
 	const { search } = useLocation();
 	const sectionName = useSelector( getSectionName );
 	const params = new URLSearchParams( search );
 	const mode = params.get( 'mode' ) as Mode;
-	const overflow = params.get( 'overflow' ) === 'true';
-	const history = useHistory();
 	const [ hideSiteInfo, setHideSiteInfo ] = useState( false );
-	const [ hasSubmittingError, setHasSubmittingError ] = useState< boolean >( false );
 	const locale = useLocale();
-	const { isLoading: submittingTicket, mutateAsync: submitTicket } = useSubmitTicketMutation();
-	const { isLoading: submittingTopic, mutateAsync: submitTopic } = useSubmitForumsMutation();
 	const userId = useSelector( getCurrentUserId );
 	const { data: userSites } = useUserSites( userId );
 	const userWithNoSites = userSites?.sites.length === 0;
-	const queryClient = useQueryClient();
-	const email = useSelector( getCurrentUserEmail );
 	const [ sitePickerChoice, setSitePickerChoice ] = useState< 'CURRENT_SITE' | 'OTHER_SITE' >(
 		'CURRENT_SITE'
 	);
@@ -174,14 +96,8 @@ export const HelpCenterContactForm = () => {
 		};
 	} );
 
-	const {
-		setSite,
-		resetStore,
-		setUserDeclaredSiteUrl,
-		setUserDeclaredSite,
-		setSubject,
-		setMessage,
-	} = useDispatch( HELP_CENTER_STORE );
+	const { setUserDeclaredSite, setSubject, setMessage } = useDispatch( HELP_CENTER_STORE );
+	const formTitles = useContactFormTitle( mode );
 
 	useEffect( () => {
 		const supportVariation = getSupportVariationFromMode( mode );
@@ -199,17 +115,12 @@ export const HelpCenterContactForm = () => {
 		}
 	}, [ userWithNoSites ] );
 
-	const formTitles = useFormTitles( mode );
-
 	let ownershipResult: AnalysisReport = useSiteAnalysis(
 		// pass user email as query cache key
 		userId,
 		userDeclaredSiteUrl,
 		sitePickerChoice === 'OTHER_SITE'
 	);
-
-	const ownershipStatusLoading = ownershipResult?.result === 'LOADING';
-	const isSubmitting = submittingTicket || submittingTopic;
 
 	// if the user picked a site from the picker, we don't need to analyze the ownership
 	if ( currentSite && sitePickerChoice === 'CURRENT_SITE' ) {
@@ -246,119 +157,9 @@ export const HelpCenterContactForm = () => {
 	);
 
 	const showingSibylResults = params.get( 'show-results' ) === 'true';
-	function handleCTA() {
-		if ( ! showingSibylResults && sibylArticles && sibylArticles.length > 0 ) {
-			params.set( 'show-results', 'true' );
-			history.push( {
-				pathname: '/contact-form',
-				search: params.toString(),
-			} );
-			return;
-		}
-		const productSlug = ( supportSite as HelpCenterSite )?.plan.product_slug;
-		const plan = getPlan( productSlug );
-		const productId = plan?.getProductId();
-		const productName = plan?.getTitle();
-		const productTerm = getPlanTermLabel( productSlug, ( text ) => text );
-
-		switch ( mode ) {
-			case 'CHAT':
-				if ( supportSite ) {
-					recordTracksEvent( 'calypso_inlinehelp_contact_submit', {
-						support_variation: 'happychat',
-						force_site_id: true,
-						location: 'help-center',
-						section: sectionName,
-					} );
-
-					recordTracksEvent( 'calypso_help_live_chat_begin', {
-						site_plan_product_id: productId,
-						is_automated_transfer: supportSite.is_wpcom_atomic,
-						force_site_id: true,
-						location: 'help-center',
-						section: sectionName,
-					} );
-					history.push( '/inline-chat' );
-					break;
-				}
-				break;
-
-			case 'EMAIL':
-				if ( supportSite ) {
-					const ticketMeta = [
-						'Site I need help with: ' + supportSite.URL,
-						`Plan: ${ productId } - ${ productName } (${ productTerm })`,
-					];
-
-					if ( getQueryArgs()?.ref === 'woocommerce-com' ) {
-						ticketMeta.push(
-							`Created during store setup on ${
-								isWcMobileApp() ? 'Woo mobile app' : 'Woo browser'
-							}`
-						);
-					}
-
-					const kayakoMessage = [ ...ticketMeta, '\n', message ].join( '\n' );
-
-					submitTicket( {
-						subject: subject ?? '',
-						message: kayakoMessage,
-						locale,
-						client: 'browser:help-center',
-						is_chat_overflow: overflow,
-						source: 'source_wpcom_help_center',
-						blog_url: supportSite.URL,
-					} )
-						.then( () => {
-							recordTracksEvent( 'calypso_inlinehelp_contact_submit', {
-								support_variation: 'kayako',
-								force_site_id: true,
-								location: 'help-center',
-								section: sectionName,
-							} );
-							history.push( '/success' );
-							resetStore();
-							// reset support-history cache
-							setTimeout( () => {
-								// wait 30 seconds until support-history endpoint actually updates
-								// yup, it takes that long (tried 5, and 10)
-								queryClient.invalidateQueries( [ `activeSupportTickets`, email ] );
-							}, 30000 );
-						} )
-						.catch( () => {
-							setHasSubmittingError( true );
-						} );
-				}
-				break;
-
-			case 'FORUM':
-				submitTopic( {
-					ownershipResult,
-					message: message ?? '',
-					subject: subject ?? '',
-					locale,
-					hideInfo: hideSiteInfo,
-					userDeclaredSiteUrl,
-				} )
-					.then( ( response ) => {
-						recordTracksEvent( 'calypso_inlinehelp_contact_submit', {
-							support_variation: 'forums',
-							force_site_id: true,
-							location: 'help-center',
-							section: sectionName,
-						} );
-						history.push( `/success?forumTopic=${ encodeURIComponent( response.topic_URL ) }` );
-						resetStore();
-					} )
-					.catch( () => {
-						setHasSubmittingError( true );
-					} );
-				break;
-		}
-	}
 
 	const InfoTip = () => {
-		const ref = useRef< any >();
+		const ref = useRef< HTMLButtonElement >( null );
 		const [ isOpen, setOpen ] = useState( false );
 
 		return (
@@ -391,20 +192,14 @@ export const HelpCenterContactForm = () => {
 
 	const shouldShowHelpLanguagePrompt = getSupportedLanguages( mode, locale );
 
-	const isCTADisabled = () => {
-		if ( isSubmitting || ! message || ownershipStatusLoading ) {
-			return true;
-		}
-
-		switch ( mode ) {
-			case 'CHAT':
-				return ! supportSite;
-			case 'EMAIL':
-				return ! supportSite || ! subject;
-			case 'FORUM':
-				return ! subject;
-		}
-	};
+	const { handleCTA, isCTADisabled, isSubmitting, hasSubmittingError } = useContactFormCTA(
+		mode,
+		supportSite as HelpCenterSite,
+		sectionName,
+		ownershipResult,
+		hideSiteInfo,
+		params
+	);
 
 	const getCTALabel = () => {
 		if ( ! showingSibylResults && sibylArticles && sibylArticles.length > 0 ) {
@@ -419,7 +214,7 @@ export const HelpCenterContactForm = () => {
 			return __( 'Still email us', __i18n_text_domain__ );
 		}
 
-		if ( ownershipStatusLoading ) {
+		if ( ownershipResult?.result === 'LOADING' ) {
 			return formTitles.buttonLoadingLabel;
 		}
 
@@ -456,11 +251,6 @@ export const HelpCenterContactForm = () => {
 		<main className="help-center-contact-form">
 			<BackButton />
 			<h1 className="help-center-contact-form__site-picker-title">{ formTitles.formTitle }</h1>
-			{ formTitles.formSubtitle && (
-				<p className="help-center-contact-form__site-picker-form-subtitle">
-					{ formTitles.formSubtitle }
-				</p>
-			) }
 
 			{ formTitles.formDisclaimer && (
 				<p className="help-center-contact-form__site-picker-form-warning">
@@ -469,32 +259,14 @@ export const HelpCenterContactForm = () => {
 			) }
 
 			{ ! userWithNoSites && (
-				<section>
-					<HelpCenterSitePicker
-						enabled={ mode === 'FORUM' }
-						currentSite={ currentSite }
-						onSelect={ ( id: string | number ) => {
-							if ( id !== 0 ) {
-								setSite( currentSite );
-							}
-							setSitePickerChoice( id === 0 ? 'OTHER_SITE' : 'CURRENT_SITE' );
-						} }
-						siteId={ sitePickerChoice === 'CURRENT_SITE' ? currentSite?.ID : 0 }
-					/>
-				</section>
-			) }
-
-			{ sitePickerChoice === 'OTHER_SITE' && (
-				<>
-					<section>
-						<TextControl
-							label={ __( 'Site address', __i18n_text_domain__ ) }
-							value={ userDeclaredSiteUrl ?? '' }
-							onChange={ setUserDeclaredSiteUrl }
-						/>
-					</section>
-					<HelpCenterOwnershipNotice ownershipResult={ ownershipResult } />
-				</>
+				<HelpCenterSitePicker
+					ownershipResult={ ownershipResult }
+					sitePickerChoice={ sitePickerChoice }
+					setSitePickerChoice={ setSitePickerChoice }
+					currentSite={ currentSite }
+					siteId={ sitePickerChoice === 'CURRENT_SITE' ? currentSite?.ID : 0 }
+					enabled={ mode === 'FORUM' }
+				/>
 			) }
 
 			{ [ 'FORUM', 'EMAIL' ].includes( mode ) && (
