+++ packages/help-center/src/components/index.d.ts
@@ -145,5 +145,3 @@ declare module 'calypso/state/analytics/actions' {
 declare module '@automattic/state-utils' {
 	export const createSelector = unknown;
 }
-
-declare module 'calypso/lib/mobile-app';
