+++ packages/help-center/src/hooks/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'calypso/lib/mobile-app';
